clean: Factor out storing local namespace in generators
The function jscode_namespace returns javascript code to either store
the namespace when leaving a generator in yield (action === 'store') or
code to restore it when resuming the generator (action === 'restore').

One reason for factoring it out is that the generated code looks the
same when storing and restoring. Before the change, the code looked
different due to the order of some of the assignments being swapped and
it was not clear at first glance what the difference between storing and
restoring is. This should now be clear. I came across this when I was
trying to understand how generators work.